+++ packages/@vue/cli-ui/src/mixins/OnWindowResize.js
@@ -1,0 +1,13 @@
+export default function () {
+  // @vue/component
+  return {
+    mounted () {
+      window.addEventListener('resize', this.onWindowResize)
+      this.onWindowResize()
+    },
+
+    beforeDestroy () {
+      window.removeEventListener('resize', this.onWindowResize)
+    }
+  }
+}